Fix #2692
