+++ src/main/java/com/fasterxml/jackson/databind/deser/BeanDeserializer.java
@@ -480,7 +480,7 @@ public class BeanDeserializer
                 // 12-Aug-2025, tatu: [databind#5237] Mergeable properties need
                 //    special handling: must defer deserialization until POJO
                 //    is constructed.
-                if (prop instanceof MergingSettableBeanProperty) {
+                if (prop.isMerging()) {
                     TokenBuffer tb = ctxt.bufferForInputBuffering(p);
                     tb.copyCurrentStructure(p);
                     buffer.bufferMergingProperty(prop, tb);
